Remove Spotlight::Dor::Resources / Spotlight::Resources namespaces from imported code
