+++ config/locales/en.yml
@@ -21,3 +21,15 @@
 
 en:
   hello: "Hello world"
+  activerecord:
+    attributes:
+      dor_harvester:
+        druid_list: Object IDs
+    help:
+      dor_harvester:
+        druid_list: Collection object PURLs, one per line (e.g. qb438pg7646)
+
+  dor_harvester:
+    form:
+      add_item: Add items
+      title: SDR
