--- config/locales/spotlight-dor-resources.en.yml
@@ -1,35 +1,0 @@
-# Files in the config/locales directory are used for internationalization
-# and are automatically loaded by Rails. If you want to use locales other
-# than English, add the necessary files in this directory.
-#
-# To use the locales, use `I18n.t`:
-#
-#     I18n.t 'hello'
-#
-# In views, this is aliased to just `t`:
-#
-#     <%= t('hello') %>
-#
-# To use a different locale, set it with `I18n.locale`:
-#
-#     I18n.locale = :es
-#
-# This would use the information in config/locales/es.yml.
-#
-# To learn more, please read the Rails Internationalization guide
-# available at http://guides.rubyonrails.org/i18n.html.
-
-en:
-  activerecord:
-    attributes:
-      spotlight/resources/dor_harvester:
-        druid_list: Object IDs
-    help:
-      spotlight/resources/dor_harvester:
-        druid_list: Collection object PURLs, one per line (e.g. qb438pg7646)
-  spotlight:
-    resources:
-      dor_harvester:
-        form:
-          add_item: Add items
-          title: SDR
