+++ db/migrate/20160728165423_change_dor_harvester_resource.rb
@@ -1,0 +1,9 @@
+class ChangeDorHarvesterResource < ActiveRecord::Migration
+  def up
+    Spotlight::Resource.where(type: 'Spotlight::Resources::DorHarvester').update_all(type: 'DorHarvester')
+  end
+
+  def down
+    Spotlight::Resource.where(type: 'DorHarvester').update_all(type: 'Spotlight::Resources::DorHarvester')
+  end
+end
