+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160712235353) do
+ActiveRecord::Schema.define(version: 20160728165423) do
 
   create_table "bookmarks", force: :cascade do |t|
     t.integer  "user_id",       null: false
